Minor caching fixes.

+++ Core/NES/BaseNesPpu.h
@@ -90,7 +90,7 @@ protected:
 	static constexpr int SpriteShifterDone = 0x8000;
 	uint16_t _spriteShifterList[9] = { SpriteShifterDone, SpriteShifterDone, SpriteShifterDone, SpriteShifterDone, SpriteShifterDone, SpriteShifterDone, SpriteShifterDone, SpriteShifterDone, SpriteShifterDone }; //Ordered by X coordinate.
 	uint8_t _nextSpriteShifter = 0;
-	uint8_t _nextSpriteShifterCycle = 0;
+	uint16_t _nextSpriteShifterCycle = 0;
 	uint8_t _activeSpriteShifters = 0;
 	uint8_t _dotSkipped = false;
 	bool _processSprites = false;
